AUTH: Disable token refresh for Planner testing
When running end-to-end tests for fabric8-planner, the token refresh happens repeatedly because of the malformed token refresh url.
The url is malformed because standalone planner does not communicate with auth. Because of the repeated requests, angular NgZone doesn't stablize and hence protractor doesn't start the tests. This patch fixes this issue.